Add condition to only intercept requests from gardenlets and service accounts from seed namespaces
This only works for 1.30+ clusters using the structured authz approach,
but it's a nice feature that we can use now (with < 1.30, it was not
possible to configure such conditions).

Also, this allows us to set the failure policy to "deny" (i.e., when our
webhook server denies a request from a gardenlet or an extension, nobody
else should be able to allow it (seed authorizers is the only instance
which may decide (except for manually configured RBAC rules)).